Files: FS_ListFilteredFiles always strips trailing slash if extension set to slash
Previously, if FS_ListFilteredFiles was called with extension set to slash it would return the path with a trailing slash if it was inside a pk3, and without a trailing slash outside.

This would lead to folders being added twice to the file list, causing problems in the species code where a model would be added multiple times, leading to the menus getting stuck.

There's still some weirdness with Sys_ListFiles not generally listing folders outside pk3s, so /dir is inconsistent between inside and outside pk3s but it isn't causing any obvious issues.